chore: refactor messages gas computation for bouncer
